fix(rosterbuilder): fix logic to determine difference in days between two dates
The previous logic returned the number of days in the month rather than the number of days total,
meaning a calculation of `dateInDays1 - dateInDays2` would break for dates in different months. This
change uses the Moment library's `diff()` function to accurately calculate the difference in days.

This commit also updates all types & interfaces using dates to use the `Date` type over string, ensuring the conversion from standard input is done correctly when running the application.