style(prettier): use pre-existing style preferences from original code
One of my first commits to this project formatted the src code based on prettier's default
configuration. This causes quite an ugly diff compared to the original code, so this commit updates
prettier to use the original style preferences and updates the src files accordingly, making for an easier code review experience.