fix(rosterbuilder): increment by 1 day only
Previously, the `build()` function would increment the day to generate shifts by `n`, which became
an incfreasingly large number due to how the array being iterated over was set up. This change
ensures we only add one day to the stateful `start` variable after we generate the shifts for that
day.