test(rosterbuilder): fix tests ensuring nurses are unique per shift and per day
Tests were missing to ensure that each shift had unique nurses on it, and I noticed the test
ensuring nurses did not work repeated shifts per day had no assertions in it. This commit adds a
test for `createShift` to ensure unique nurses, and fixes the `createDay` test to assert that nurses
did not work twice per day.